refactor: colocate @react-email/code-block source in react-email
Move code-block implementation under packages/react-email/src/components/code-block.
Add prismjs to react-email dependencies. Point @react-email/code-block build at
the shared entry with prismjs external. Use PrismToken alias for strict typing.

Made-with: Cursor
