--- packages/code-block/src/index.ts
@@ -1,3 +1,0 @@
-export * from './code-block';
-export * from './languages-available';
-export * from './themes';
