+++ packages/react-email/src/components/code-block/index.ts
@@ -1,0 +1,3 @@
+export * from './code-block.js';
+export * from './languages-available.js';
+export * from './themes.js';
