+++ packages/react-email/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './body/index.js';
 export * from './button/index.js';
+export * from './code-block/index.js';
